feat(gui): air out the plate geometry on the owner's device round
The shared plate takes 4 pt extra under the card's top edge —
every tile sat visibly close to it. Gaps & Borders centres an
implied 16:10 screen outline and reads the outer gaps as
insets from THAT edge instead of stretching to the plate's.
The Spaces fan's panes become 16:10 landscape screens (70x44,
18 pt exposure), centred on both axes so the leading and
trailing margins match. Advanced Colours' swatches go
landscape 2:1 (46x23), spread with breathing room on every
side. The legibility floor splits to its own file at the
350-line ceiling.
