+++ Sources/KiwiDesk/Settings/HomeCardPlate.swift
@@ -25,9 +25,10 @@ enum HomeCardPlate {
                 HomeCardSpacesTile(model: model)
             }
         case .gapsAndBorders:
-            // Padding 4, not the family 11: the tile's own
-            // per-edge insets ARE the outer-gap readout, so a
-            // wide fixed frame would drown the value it shows.
+            // Padding 4, not the family 11: the tile centres
+            // its own 16:10 screen and the outer-gap readouts
+            // inset from THAT outline, so the plate's frame
+            // stays thin around it.
             return tile(padding: 4, settings: settings) {
                 HomeCardGapsTile(settings: settings)
             }
@@ -95,6 +96,11 @@ enum HomeCardPlate {
                     maxHeight: .infinity
                 )
                 .padding(padding)
+                // Every picture takes extra air under the
+                // card's top edge — at the uniform inset alone
+                // the tiles sat visibly close to it (owner,
+                // 2026-08-09, on device).
+                .padding(.top, 4)
                 .frame(height: SettingsTheme.plateHeight)
                 .frame(maxWidth: .infinity)
                 .background(SettingsTheme.previewPlate)
@@ -134,53 +140,6 @@ enum HomeCardPlate {
             base: SettingsTheme.previewPlate
         )
     }
-
-    /// The plate's own weighted luminance — `previewPlate`'s
-    /// `0x12251A`, restated as arithmetic so the floor below
-    /// can composite translucent user colours against it.
-    private static let plateLuminance =
-        0.2126 * (0x12 / 255.0)
-        + 0.7152 * (0x25 / 255.0)
-        + 0.0722 * (0x1A / 255.0)
-
-    /// Whether a user hex can be seen on the plate at all: its
-    /// alpha-composited weighted luminance must sit a floor's
-    /// width from the plate's own, in either direction — a
-    /// near-plate dark and a translucent wisp both fail.
-    /// Internal so `HomeCardChromeTests` can pin the floor with
-    /// the palettes that motivated it.
-    static func plateLegible(_ hex: String) -> Bool {
-        guard let c = rgba(hex) else { return false }
-        let lum =
-            0.2126 * c.r + 0.7152 * c.g + 0.0722 * c.b
-        let effective =
-            c.a * lum + (1 - c.a) * plateLuminance
-        return abs(effective - plateLuminance) >= 0.15
-    }
-
-    private static func rgba(
-        _ hex: String
-    ) -> (r: Double, g: Double, b: Double, a: Double)? {
-        var s = hex.trimmingCharacters(in: .whitespaces)
-        if s.hasPrefix("#") { s.removeFirst() }
-        guard s.count == 6 || s.count == 8,
-            let v = UInt64(s, radix: 16)
-        else { return nil }
-        if s.count == 6 {
-            return (
-                Double((v >> 16) & 0xFF) / 255,
-                Double((v >> 8) & 0xFF) / 255,
-                Double(v & 0xFF) / 255,
-                1
-            )
-        }
-        return (
-            Double((v >> 24) & 0xFF) / 255,
-            Double((v >> 16) & 0xFF) / 255,
-            Double((v >> 8) & 0xFF) / 255,
-            Double(v & 0xFF) / 255
-        )
-    }
 }
 
 /// The Spaces & Layouts picture: a FAN of mini-desktops, one
@@ -199,8 +158,12 @@ struct HomeCardSpacesTile: View {
     @ObservedObject var model: SettingsModel
     @Environment(\.schematicPalette) private var palette
 
-    private static let card = CGSize(width: 34, height: 56)
-    private static let exposed: CGFloat = 16
+    // 16:10 landscape panes, not tall slabs — each pane is a
+    // SCREEN (owner, 2026-08-09) — centred on both axes so the
+    // first pane's leading margin equals the last one's
+    // trailing margin.
+    private static let card = CGSize(width: 70, height: 44)
+    private static let exposed: CGFloat = 18
     private static let cap = 8
 
     var body: some View {
